Add bcrypt as a dependency
CI was failing because the generated user model was being placed in the test directory as expected. I think the problem was that the test command was testing that file before it could be deleted. Since the user model contains a call to `has_secure_password` which requires bcrypt, CI was raising an error.

I think the larger the install command is, the slower the test suite will run, which will result in intermittent failures. I was not running into this locally because my machine simply has more resources.